🚀 Add donation item

+++ app/src/main/java/com/mashup/mobalmobal/ui/profile/domain/model/ProfileItem.kt
@@ -37,8 +37,8 @@ sealed class ProfileItem(
         val description: String,
         val goalPrice: Double,
         val donatedPrice: Double,
-        val startDate: Double,
-        val dueDate: Double,
+        val startDate: Long,
+        val dueDate: Long,
     ) : ProfileItem(donationId)
 
     data class Donated(
@@ -48,7 +48,7 @@ sealed class ProfileItem(
         val description: String,
         val goalPrice: Double,
         val donatedPrice: Double,
-        val startDate: Double,
-        val dueDate: Double,
+        val startDate: Long,
+        val dueDate: Long,
     ) : ProfileItem(donationId)
 }
